backend/requirements.txt
fastapi[standard]
schemathesis
pytest
uvicorn
python-dotenv
pytest-cov
httpx
pydantic[email]